restructured library directory structire
