+++ library/lib/EasyBMP/EasyBMP.cpp
