+++ library/lib/EasyBMP/EasyBMP.h
