+++ library/lib/EasyBMP/EasyBMP_DataStructures.h
