+++ library/lib/EasyBMP/EasyBMP_VariousBMPutilities.h
