+++ library/lib/lodepng/lodepng.h
