+++ test/CMakeLists.txt
@@ -4,6 +4,9 @@ project(wfcpp-test VERSION 0.0.1 LANGUAGES CXX)
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
+#wcfpp-test.exe
+add_executable(main.exe src/main.cpp)
+
 if (CMAKE_BUILD_TYPE MATCHES "^[Dd]ebug")
   # set compiler flags
   set(CMAKE_CXX_FLAGS "-O0 -coverage -Wall")
@@ -20,13 +23,14 @@ if (CMAKE_BUILD_TYPE MATCHES "^[Dd]ebug")
     COMMAND ${GENHTML} --demangle-cpp -o coverage ../coverage.info
     )
 
+  target_link_libraries(main.exe wfcpp -coverage)
+
 else()
   # set compiler flags
-  set(CMAKE_CXX_FLAGS "-O0 -Wall")
+  set(CMAKE_CXX_FLAGS "-O2")
+  target_link_libraries(main.exe wfcpp)
 endif()
 
-#wcfpp-test.exe
-add_executable(main.exe src/main.cpp)
 
 target_link_libraries(main.exe wfcpp)
